Side menu and records
Added a functional side menu to increase navigation options and began implementation of a records system, which would allow users to locally record and store information about their workouts, such as what type (lifting vs. cardio, possible other in future, like sports), as well as specific information depending on the type chosen, such as workout length and what specifically was done during the workout, and records for the date and time. Side menu navigates to these pages, and is the only current way to navigate to these pages